FDG-9789 Move props to context
