+++ src/components/data-preview/data-preview-context.jsx
@@ -1,6 +1,7 @@
 import React, { createContext, useEffect, useState } from 'react';
 import { columnsConstructorData } from '../data-table/data-table-helper';
 import GLOBALS from '../../helpers/constants';
+import { getCoreRowModel, getFilteredRowModel, getPaginationRowModel, getSortedRowModel, useReactTable } from '@tanstack/react-table';
 
 const DEFAULT_ROWS_PER_PAGE = GLOBALS.dataTable.DEFAULT_ROWS_PER_PAGE;
 
@@ -18,10 +19,16 @@ const DataTableProvider = ({ children, config, detailViewState, pivotSelected })
   const [maxPage, setMaxPage] = useState(1);
   const [maxRows, setMaxRows] = useState(1);
   const [currentPage, setCurrentPage] = useState(1);
+  const [manualPagination, setManualPagination] = useState(false);
+  const [reactTableSorting, setReactTableSort] = useState([]);
 
   //dependent on table props
   const [itemsPerPage, setItemsPerPage] = useState();
   const [rowsShowing, setRowsShowing] = useState({ begin: 1, end: 1 });
+
+  // const defaultSelectedColumns = config?.detailView?.selectColumns && detailViewState ? config.detailView.selectColumns : tableProps?.selectColumns;
+
+  const [columnVisibility, setColumnVisibility] = useState();
 
   const handlePerPageChange = numRows => {
     const numItems = numRows >= maxRows ? maxRows : numRows;
@@ -78,31 +85,48 @@ const DataTableProvider = ({ children, config, detailViewState, pivotSelected })
           ? tableProps?.data?.length
           : DEFAULT_ROWS_PER_PAGE
       );
+      const defaultSelectedColumns =
+        config?.detailView?.selectColumns && detailViewState ? config.detailView.selectColumns : tableProps.selectColumns;
+      const defaultInvisibleColumns = {};
+      // We need to be able to access the accessorKey (which is a type violation) hence the ts ignore
+      if (defaultSelectedColumns && allColumns) {
+        for (const column of allColumns) {
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+          // @ts-ignore
+          if (defaultSelectedColumns && !defaultSelectedColumns?.includes(column.accessorKey)) {
+            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+            // @ts-ignore
+            defaultInvisibleColumns[column.accessorKey] = false;
+          }
+        }
+      }
+
+      setColumnVisibility(defaultSelectedColumns && defaultSelectedColumns.length > 0 && !pivotSelected ? defaultInvisibleColumns : {});
     }
   }, [tableProps]);
 
-  // const table = useReactTable({
-  //   columns: allColumns,
-  //   data: reactTableData.data,
-  //   columnResizeMode: 'onChange',
-  //   initialState: {
-  //     pagination: {
-  //       pageIndex: 0,
-  //       pageSize: pagingProps.itemsPerPage,
-  //     },
-  //   },
-  //   state: {
-  //     columnVisibility,
-  //     sorting,
-  //   },
-  //   onSortingChange: setSorting,
-  //   onColumnVisibilityChange: setColumnVisibility,
-  //   getCoreRowModel: getCoreRowModel(),
-  //   getPaginationRowModel: getPaginationRowModel(),
-  //   getSortedRowModel: getSortedRowModel(),
-  //   getFilteredRowModel: getFilteredRowModel(),
-  //   manualPagination: manualPagination,
-  // });
+  const table = useReactTable({
+    columns: allColumns,
+    data: reactTableData?.data,
+    columnResizeMode: 'onChange',
+    initialState: {
+      pagination: {
+        pageIndex: 0,
+        pageSize: pagingProps.itemsPerPage,
+      },
+    },
+    state: {
+      columnVisibility,
+      sorting: reactTableSorting,
+    },
+    onSortingChange: setReactTableSort,
+    onColumnVisibilityChange: setColumnVisibility,
+    getCoreRowModel: getCoreRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    manualPagination: manualPagination,
+  });
   return (
     <DataTableContext.Provider
       value={{
@@ -131,7 +155,14 @@ const DataTableProvider = ({ children, config, detailViewState, pivotSelected })
         rowsShowing,
         setRowsShowing,
         data,
-        // table,
+        manualPagination,
+        setManualPagination,
+        reactTableSorting,
+        setReactTableSort,
+        columnVisibility,
+        setColumnVisibility,
+        table,
+        // defaultSelectedColumns,
       }}
     >
       {children}
